Bugfix
Using the wrong key to get the value out of the database when resolving
config variables.
The only reason this worked was because some of the maps had the same
name for both key and value (I'm not forcing them to have different
values so that problems like this are caught sooner).