Chalk Compatibility and Hardware info
There is some additional formatting done to the cards that are displayed
if the app is used on a Pebble Time Round (chalk). This was done without
cloudpebble (since that doesn't support Chalk yet). This should be
considered a temporary stopgap until official support is added

The watch now also sends hardware information to the server to aid with
debugging and tracking down bugs